refactor(user): update getAllUser method to return ReturnUserDTO
Refactored the getAllUser method in order to return an array of ReturnUserDTO objects instead of
UserEntity objects. This change involved importing the UserEntity class and creating a new
ReturnUserDTO class. The ReturnUserDTO class has properties for id, name, email, cpf, and phone,
which are assigned values from the corresponding properties in the UserEntity object.