feat(category): add CategoryEntity and categoryMock
The CategoryEntity class defines the structure of the "category" entity in the database. It includes
columns such as id, name, createdAt, and updatedAt. It also establishes a one-to-many relationship
with the ProductEntity class.The categoryMock object is a mock data instance of the CategoryEntity
class. It includes sample values for createdAt, updatedAt, id, and name.