feat(category): add ReturnCategory, CreateCategory, and createCategoryMock
The ReturnCategory class is used to define the structure of the returned category object. It
includes id and name properties. The constructor accepts an object of type CategoryEntity and
assigns the values of id and name from it to the class properties.The CreateCategory class is a DTO
(Data Transfer Object) used for creating a new category. It includes a single property, name, which
is decorated with the @IsString() validator to ensure it is a string.The createCategoryMock object
is a mock data instance of the CreateCategory class. It includes a sample value for the name
property.