feat(cart): add CartController and CartService
The CartController utilizes the InsertCartDTO to receive cart creation data from the request body.
It also applies a ValidationPipe to validate the incoming data. The createCart method in the
controller delegates the responsibility of inserting the product into the cart to the CartService.
The CartService class is responsible for handling cart-related operations. It leverages the
CartEntity and InsertCartDTO to interact with the database. The verifyActiveCart method checks if
there is an active cart for the specified user and throws a NotFoundException if not found. The
createCart method creates a new cart for the user. The insertProductInCart method validates the
user's active cart and inserts the product into the cart using the CartProductService.