feat(cart): add deleteProductCart and updateProductInCart functions
Added deleteProductCart function to delete a product from the user's cart. It retrieves the cart ID
based on the user ID and then calls the deleteProductCart function from the cartProductService. add
updateProductInCart function to handle updating the cart with a new product. It first tries to find
the cart based on the user ID. If it fails, it creates a new cart for the user. Then, it calls the
updateProductInCart function from the cartProductService.