test(cart): add unit tests for cartService methods
The findCartByUserId method is tested to ensure it returns the cart successfully both with and
without sending relations. Additionally, it tests if the method throws a NotFoundException when the
cart is not found.The createCart method is tested to verify if it saves the cart with the correct
information. For the insertProductInCart method, the tests check if it returns "cart not found" when
the cart is not found, and "cart found" when the cart is present.