refactor(payment): optimize finalPrice calculation in PaymentService
This commit refactors the createPayment method in the PaymentService class, extracting the logic for
calculating the finalPrice into a separate method called generateFinalPrice. This makes the code
more modular and easier to maintain. Additionally, it optimizes the finalPrice calculation for
different payment scenarios (credit card and pix) based on the createOrderDTO data