refactor(product): getAllProduct method with optional relation retrieval
The getAllProduct method by introducing an optional parameter isFindRelations, which allows the
method to retrieve related data if set to true. This is an enhancement to the original method, which
only returned the product entities based on the given productId. Now, with the added flexibility,
the method can also fetch related data (in this case, the category relation) by setting
isFindRelations to true.