feat(product): add findPriceDelivery endpoint and ReturnPriceDeliveryDTO
- Implemented the 'findPriceDelivery' endpoint with the corresponding route and parameters in the
ProductService.
- The endpoint fetches the product details and creates a SizeProductDTO with the
product's dimensions and value.
- It then calls the 'priceDelivery' method of the CorreiosService
to obtain delivery prices for different types of postal services (PAC, SEDEX, and SEDEX_10) based on
the provided CEP and product details.
- The results are aggregated and processed into the
ReturnPriceDeliveryDTO class to format and return the delivery information in a standardized way.
-
The ReturnPriceDeliveryDTO class filters and maps the response from the CorreiosService to extract
relevant delivery prices, times, and delivery type (code) for further processing.