refactor: findProductById method with optional relations parameter
Updated the findProductById method in the ProductService to accept an additional optional parameter
"isRelations" of type boolean. If "isRelations" is provided and set to true, the method will include
the "category" relation in the query to fetch category details along with the product. If
"isRelations" is not provided or set to false, the method will only fetch the product entity without
related entities.