refactor: refactor findProductById test cases for optional relations
Added a new test case to check if the method makes the correct call to the productRepository findOne
method when "isRelations" is true. In the new test case, a spy is used to mock the productRepository
findOne method, and then the findProductById method is called with the "isRelations" parameter set
to true. The test verifies that the productRepository findOne method is called with the expected
parameters, including the "relations" parameter set to { category: true }.